Only remove existing files on error

+++ trace_for_guess/concatenate.py
@@ -36,8 +36,9 @@ def cat_files(filelist, out_file):
         subprocess.run(['ncrcat'] + filelist + [out_file], check=True)
     except Exception:
         for f in [out_file] + glob.glob(out_file + 'pid*.ncrcat.tmp'):
-            cprint(f"Removing file '{f}'.", 'red')
-            os.remove(out_file)
+            if os.path.isfile(out_file):
+                cprint(f"Removing file '{f}'.", 'red')
+                os.remove(out_file)
         raise
     if not os.path.isfile(out_file):
         raise RuntimeError('The command `ncrcat` didn’t produce an output '
